Split chart generation into separate file
